:lipstick: feat and refactor: css for editUser
