+++ src/assets/css/postForm.css
@@ -1,7 +1,13 @@
 @import "globalVariables.css";
 
 .postForm-button-pink {
-  background-color: var(--button-primary-color);
+  background: var(--button-primary-color);
+  border: none;
+  border-radius: 4px;
+  padding: 11px;
+  box-shadow: none;
+  margin-top: 35px;
+  text-shadow: none;
 }
 
 .postForm-text-input {
@@ -27,3 +33,13 @@
   text-indent: 6px;
   height: 40px;
 }
+
+/* .signup-button {
+  background: var(--button-primary-color);
+  border: none;
+  border-radius: 4px;
+  padding: 11px;
+  box-shadow: none;
+  margin-top: 35px;
+  text-shadow: none;
+} */
