+++ src/assets/css/user/editUser.css
@@ -1,0 +1,20 @@
+.editUser-button-pink {
+  background: var(--button-primary-color);
+  border: none;
+  border-radius: 4px;
+  padding: 11px;
+  box-shadow: none;
+  text-shadow: none;
+}
+
+.editUser-text-input {
+  background: var(--background-secondary);
+  border: none;
+  border-bottom: 1px solid var(--border-primary);
+  border-radius: 0;
+  box-shadow: none;
+  outline: none;
+  color: inherit;
+  text-indent: 6px;
+  height: 40px;
+}
